// apps/web/src/api/generated/schemas/schemasInvoicesNewInvoiceStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */

/**
 * Invoice status enumeration
 */
export type SchemasInvoicesNewInvoiceStatus = typeof SchemasInvoicesNewInvoiceStatus[keyof typeof SchemasInvoicesNewInvoiceStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SchemasInvoicesNewInvoiceStatus = {
  pending: 'pending',
  RECEIVED: 'RECEIVED',
  PROCESSED: 'PROCESSED',
  rejected: 'rejected',
  draft: 'draft',
  SENT: 'SENT',
  paid: 'paid',
  overdue: 'overdue',
} as const;